Prepare for trouble! And make it double!

+++ src/org/usfirst/frc/team3695/robot/subsystems/SubsystemDrive.java
@@ -153,7 +153,7 @@ public class SubsystemDrive extends Subsystem {
 		for(double d : list) {
 			sum += d;
 		}
-		return sum / list.length;
+		return sum / (double)list.length;
 	}
 }
